requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
python-telegram-bot==21.7
redis==5.2.0
sniffio==1.3.1
